requirements.txt
tensorflow
djangorestframework
insightface
faiss-cpu
imutils
torch
websockets
keras
opencv-python
django-filter
psycopg2
onnxruntime-gpu
python-dotenv
celery
redis
psutil